prop declaration now in same files
