+++ src/components/ProjectComponent.tsx
@@ -1,8 +1,8 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faPenToSquare } from '@fortawesome/free-solid-svg-icons'
 import { faTrash } from '@fortawesome/free-solid-svg-icons'
-import { ProjectComponentProps } from '../types'
 import React from 'react'
+import { Project } from '../types'
 
 export const ProjectComponent: React.FC<ProjectComponentProps> = ({project,deleteproject,editproject}) => {
   return (
@@ -14,4 +14,11 @@ export const ProjectComponent: React.FC<ProjectComponentProps> = ({project,delet
       </div>
     </div>
   )
+}
+
+export interface ProjectComponentProps {
+  key: string;
+  project: Project;
+  deleteproject: (id: string) => void;
+  editproject: (id: string) => void;
 }
